+++ .changeset/lazy-suns-marry.md
@@ -1,0 +1,5 @@
+---
+"@actnowcoalition/ui-components": patch
+---
+
+Add social meta tags
