+++ packages/ui-components/src/components/SocialMetaTags/index.ts
@@ -1,1 +1,3 @@
-export * from "./SocialMetaTags";
+export * from "./PageMetaTags";
+export * from "./FacebookMetaTags";
+export * from "./TwitterMetaTags";
